text_view: Incremental Markdown parsing.  (#1792)
Only for `Markdown` parse, if we use `TextViewState::push_str` method to
append content to state, will trigger the incremental parsing.

Only the last **block node** (e.g.: A paragraph, A Heading ...) will to
parse again when we called `push_str`, this will significantly reduce
CPU overhead.

## CPU Usage on parsing

> In release mode of the `stream_markdown` example.

| Before | After |
| --- | --- |
| 40% ~ 50%    | 10% ~ 15% |



https://github.com/user-attachments/assets/179f65ea-d120-4f6c-af7d-df8b5afec201

`Html` format not support, because it is a nested layout.